services/idp_stub/pyproject.toml
[project]
name = "dadi-idp-stub"
version = "0.1.0"
description = "Local OIDC IdP stub (JWKS + token issuer) for dev/testing."
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.24",
  "pydantic>=2.0",
  "cryptography>=42.0",
  "PyJWT[crypto]>=2.8",
]
